[level 0] Title: 배열의 평균값, Time: 0.18 ms, Memory: 31.3 MB -BaekjoonHub

+++ 프로그래머스/0/120817. 배열의 평균값/배열의 평균값.cs
@@ -2,12 +2,13 @@ using System;
 
 public class Solution {
     public double solution(int[] numbers) {
-        int answer = 0;
-        double l = numbers.Length;
-        
-        for (int i = 0 ; i < l; i ++) {
+        double answer = 0;
+        int l = numbers.Length;
+
+        for (int i = 0; i < l; i++) {
             answer += numbers[i];
         }
+        
         return answer / l;
     }
 }
